update upload and show image

+++ app/Http/Controllers/ShopCRUDController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Shop;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use Illuminate\Support\Facades\File;
 
 class ShopCRUDController extends Controller
 {
@@ -56,18 +58,27 @@ class ShopCRUDController extends Controller
 
     );
 
-        $path = $request->file('s_image')->store('public/images');
+
         $shop = new Shop;
+
+        if ($request->hasFile('s_image')) {
+            $path = Str::random(10) . '.' . $request->file('s_image')->getClientOriginalExtension();
+            $request->file('s_image')->move(public_path() . '/assets/images/', $path);
+            $shop->s_image = $path;
+            }
+        else {
+                $shop->s_image = 'no_img.png';
+            }
+
         $shop ->s_name = $request->s_name;
         $shop ->s_detail = $request->s_detail;
         $shop ->s_ampher = $request->s_ampher;
         $shop ->s_lat = $request->s_lat;
         $shop ->s_long = $request->s_long;
-        $shop->s_image = $path;
         $shop ->save();
 
         return redirect()->route('shops.index')
-        -> with('success','บันทึกรายการ ร้านค้า/ของฝาก สำเร็จ !!');
+        -> with('success','บันทึกรายการ ร้านค้า/ของฝาก สำเร็จ');
 
     }
 
@@ -107,7 +118,8 @@ class ShopCRUDController extends Controller
             's_name'=> 'required',
             's_detail' => 'required',
             's_lat'=> 'required',
-            's_long'=> 'required',] ,
+            's_long'=> 'required',
+            's_image' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',],
 
             [
                 's_name.required' => 'กรุณาใส่ชื่อ ร้านค้า/ของฝาก ด้วยครับ' ,
@@ -120,10 +132,11 @@ class ShopCRUDController extends Controller
 
             $shop = Shop::find($id);
             if ($request->hasFile('s_image')){
-                Storage::delete($shop->s_image);
-                $request->validate(['s_image' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
-                ]);
-                $path = $request->file('s_image')->store('public/images');
+                if ($shop->s_image != 'nopic.jpg') {
+                    File::delete(public_path() . '\\assets\\images\\' . $shop->s_image);
+                    }
+                $path = Str::random(10) . '.' . $request->file('s_image')->getClientOriginalExtension();
+                $request->file('s_image')->move(public_path() . '/assets/images/', $path);
                 $shop->s_image = $path;
             }
 
@@ -133,8 +146,7 @@ class ShopCRUDController extends Controller
             $shop ->s_lat = $request->s_lat;
             $shop ->s_long = $request->s_long;
             $shop ->save();
-
-            return redirect()->route('shops.index')->with('success','อัพเดทรายการ ร้านค้า/ของฝาก สำเร็จ !!');
+            return redirect()->route('shops.index')->with('success','อัพเดทรายการ ร้านค้า/ของฝาก สำเร็จ');
         }
     }
 
@@ -146,9 +158,12 @@ class ShopCRUDController extends Controller
      */
     public function destroy(Shop $shop)
     {
-        Storage::delete($shop->s_image);
+
+    if ($shop->s_image != 'nopic.jpg') {
+        File::delete(public_path() . '\\assets\\images\\' . $shop->s_image);
+        }
         $shop->delete();
         return redirect()->route('shops.index')
-                        ->with('success','ลบรายการ ร้านค้า/ของฝาก สำเร็จ !!');
+                        ->with('success','ลบรายการ ร้านค้า/ของฝาก สำเร็จ');
     }
 }
